[utils][m]: get server url moved to utils -refs #45
