+++ lib/get.js
@@ -1,6 +1,5 @@
 const axios = require('axios')
 const chalk = require('chalk')
-const config = require('./config')
 const { elephant } = require('./utils/logo')
 const fs = require('fs')
 const mkdirp = require('mkdirp')
@@ -54,15 +53,8 @@ module.exports.get = async(pkgid) => {
 
 }
 
-const getServerUrl = (path=config.configDir) => {
-  let conf = config.readConfig(path)
-  if (conf) {
-    return conf.server
-  }
-  return config.defaultServer
-}
 
-const getMetadata = async (publisher, package, sUrl=getServerUrl() ) => {
+const getMetadata = async (publisher, package, sUrl=utils.getServerUrl() ) => {
   let apiUrl = `${sUrl}/api/package/${publisher}/${package}`
   let res = await axios.get(apiUrl).catch( err => {
     if (err.response && err.response.status === 404) {
@@ -136,7 +128,6 @@ const downloadFile = async (bitStoreUrl, dest, publisher, pkg, bar) => {
 }
 
 
-module.exports.getServerUrl = getServerUrl
 module.exports.getMetadata = getMetadata
 module.exports.checkDestIsEmpty = checkDestIsEmpty
 module.exports.downloadFile = downloadFile
