+++ lib/utils/common.js
@@ -1,5 +1,7 @@
+const config = require('../config')
+
 const parseIdentifier = (dhpkgid) => {
-  let [publisher, pkg, path ] = dhpkgid.split('/')
+  const [publisher, pkg, path ] = dhpkgid.split('/')
   return {
     publisher,
     pkg,
@@ -7,4 +9,13 @@ const parseIdentifier = (dhpkgid) => {
   }
 }
 
+const getServerUrl = (path=config.configDir) => {
+  const conf = config.readConfig(path)
+  if (conf) {
+    return conf.server
+  }
+  return config.defaultServer
+}
+
+module.exports.getServerUrl = getServerUrl
 module.exports.parseIdentifier = parseIdentifier
